moving getserviceaccountandverifyworkloadidentity
