+++ pkg/controller/controllerutils/controller_errors.go
@@ -1,4 +1,4 @@
-package controllererrors
+package controllerutils
 
 type UserError struct {
 	Err         error
