+++ pkg/controller/keyvault/nginx_secret_provider_class.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 
 	approutingv1alpha1 "github.com/Azure/aks-app-routing-operator/api/v1alpha1"
-	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllererrors"
+	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllerutils"
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -110,7 +110,7 @@ func (i *NginxSecretProviderClassReconciler) Reconcile(ctx context.Context, req 
 		}
 		err = buildSPC(spc, spcConf)
 		if err != nil {
-			var userErr controllererrors.UserError
+			var userErr controllerutils.UserError
 			if errors.As(err, &userErr) {
 				logger.Info(fmt.Sprintf("failed to build secret provider class for nginx ingress controller with error: %s. sending warning event", userErr.Error()))
 				i.events.Eventf(nic, corev1.EventTypeWarning, "InvalidInput", "error while processing Keyvault reference: %s", userErr.UserError())
