+++ pkg/controller/keyvault/placeholder_pod.go
@@ -11,7 +11,7 @@ import (
 	"strconv"
 
 	"github.com/Azure/aks-app-routing-operator/api/v1alpha1"
-	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllererrors"
+	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllerutils"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
@@ -176,7 +176,7 @@ func (p *PlaceholderPodController) reconcileObjectDeployment(dep *appsv1.Deploym
 	// Verify ServiceAccount exists (if Gateway)
 	serviceAccount, err = p.verifyServiceAccount(ctx, spc, obj, logger)
 	if err != nil {
-		var userErr controllererrors.UserError
+		var userErr controllerutils.UserError
 		if errors.As(err, &userErr) {
 			logger.Info("user error while verifying if service account exists: %s", userErr.Err)
 			p.events.Eventf(obj, corev1.EventTypeWarning, "InvalidInput", userErr.UserMessage)
@@ -356,10 +356,10 @@ func (p *PlaceholderPodController) verifyServiceAccount(ctx context.Context, spc
 
 		if serviceAccount == "" {
 			err := fmt.Errorf("failed to locate listener for SPC %s on user's gateway resource", spc.Name)
-			return "", controllererrors.NewUserError(err, fmt.Sprintf("gateway listener for spc %s doesn't exist or doesn't contain required TLS options", spc.Name))
+			return "", controllerutils.NewUserError(err, fmt.Sprintf("gateway listener for spc %s doesn't exist or doesn't contain required TLS options", spc.Name))
 		}
 
-		_, err := GetServiceAccountAndVerifyWorkloadIdentity(ctx, p.client, serviceAccount, spc.Namespace)
+		_, err := controllerutils.GetServiceAccountAndVerifyWorkloadIdentity(ctx, p.client, serviceAccount, spc.Namespace)
 		if err != nil {
 			return "", err
 		}
