+++ pkg/controller/keyvault/placeholder_pod_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 
 	"github.com/Azure/aks-app-routing-operator/api/v1alpha1"
-	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllererrors"
 	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllername"
+	"github.com/Azure/aks-app-routing-operator/pkg/controller/controllerutils"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
@@ -536,7 +536,7 @@ func TestVerifyServiceAccount(t *testing.T) {
 			spc:             serviceAccountTwoSpc,
 			obj:             modifyGateway(gatewayWithTwoServiceAccounts, func(gw *gatewayv1.Gateway) { gw.Spec.Listeners[1].Name = "test-listener-3" }),
 			existingObjects: []client.Object{gatewayWithTwoServiceAccounts, annotatedServiceAccount},
-			expectedError:   controllererrors.NewUserError(errors.New("failed to locate listener for SPC kv-gw-cert-test-gw-test-listener-2 on user's gateway resource"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
+			expectedError:   controllerutils.NewUserError(errors.New("failed to locate listener for SPC kv-gw-cert-test-gw-test-listener-2 on user's gateway resource"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
 		},
 		{
 			name: "listener matches but doesn't contain service account option",
@@ -545,7 +545,7 @@ func TestVerifyServiceAccount(t *testing.T) {
 				gw.Spec.Listeners[1].TLS.Options = map[gatewayv1.AnnotationKey]gatewayv1.AnnotationValue{"not-service-account": "test-value"}
 			}),
 			existingObjects: []client.Object{gatewayWithTwoServiceAccounts, serviceAccountTwoSpc, annotatedServiceAccount},
-			expectedError:   controllererrors.NewUserError(errors.New("failed to locate listener for SPC kv-gw-cert-test-gw-test-listener-2 on user's gateway resource"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
+			expectedError:   controllerutils.NewUserError(errors.New("failed to locate listener for SPC kv-gw-cert-test-gw-test-listener-2 on user's gateway resource"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
 		},
 		{
 			name: "nonexistent service account referenced",
@@ -554,7 +554,7 @@ func TestVerifyServiceAccount(t *testing.T) {
 				gw.Spec.Listeners[1].TLS.Options = map[gatewayv1.AnnotationKey]gatewayv1.AnnotationValue{"kubernetes.azure.com/tls-cert-service-account": "fake-sa"}
 			}),
 			existingObjects: []client.Object{serviceAccountTwoSpc, gatewayWithTwoServiceAccounts, annotatedServiceAccount},
-			expectedError:   controllererrors.NewUserError(errors.New("serviceaccounts \"fake-sa\" not found"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
+			expectedError:   controllerutils.NewUserError(errors.New("serviceaccounts \"fake-sa\" not found"), "gateway listener for spc %s doesn't exist or doesn't contain required TLS options"),
 		},
 		{
 			name: "service account without required annotation referenced",
@@ -575,7 +575,7 @@ func TestVerifyServiceAccount(t *testing.T) {
 					},
 				},
 			},
-			expectedError: controllererrors.NewUserError(errors.New("user-specified service account does not contain WI annotation"), "serviceAccount test-sa was specified in Gateway but does not include necessary annotation for workload identity"),
+			expectedError: controllerutils.NewUserError(errors.New("user-specified service account does not contain WI annotation"), "serviceAccount test-sa was specified in Gateway but does not include necessary annotation for workload identity"),
 		},
 		{
 			name: "incorrect object type",
